requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.1.0
python-dotenv==1.0.0
flask-bcrypt==1.0.1
pillow==10.1.0